requirements.txt
Flask>=2.2
requests>=2,<3
msal>=1.20.0
python-dotenv>=0.20,<1
gunicorn>=20